Grouping diffusion solvers into a diffusion module
