--- lua/modules/cfem_diffusion/cfem_diff_solver.h
@@ -1,13 +1,0 @@
-// SPDX-FileCopyrightText: 2024 The OpenSn Authors <https://open-sn.github.io/opensn/>
-// SPDX-License-Identifier: MIT
-
-#pragma once
-
-#include "framework/parameters/input_parameters.h"
-
-namespace opensnlua
-{
-
-opensn::ParameterBlock CFEMSetOptions(const opensn::InputParameters& params);
-
-}
