--- lua/modules/mg_diffusion/mg_diff_solver.h
@@ -1,13 +1,0 @@
-// SPDX-FileCopyrightText: 2024 The OpenSn Authors <https://open-sn.github.io/opensn/>
-// SPDX-License-Identifier: MIT
-
-#pragma once
-
-#include "lua/framework/lua.h"
-
-namespace opensnlua
-{
-
-opensn::ParameterBlock CFEMMGSetOptions(const opensn::InputParameters& params);
-
-} // namespace opensnlua
