+++ modules/diffusion/cfem_diffusion_solver.cc
@@ -1,14 +1,13 @@
 // SPDX-FileCopyrightText: 2024 The OpenSn Authors <https://open-sn.github.io/opensn/>
 // SPDX-License-Identifier: MIT
 
-#include "modules/cfem_diffusion/cfem_diffusion_solver.h"
+#include "modules/diffusion/cfem_diffusion_solver.h"
 #include "framework/runtime.h"
 #include "framework/object_factory.h"
 #include "framework/logging/log.h"
 #include "framework/utils/timer.h"
 #include "framework/math/functions/scalar_spatial_material_function.h"
 #include "framework/mesh/mesh_continuum/mesh_continuum.h"
-#include "modules/common/diffusion_bndry.h"
 #include "framework/field_functions/field_function_grid_based.h"
 #include "framework/math/spatial_discretization/finite_element/piecewise_linear/piecewise_linear_continuous.h"
