+++ modules/diffusion/cfem_diffusion_solver.h
@@ -5,11 +5,11 @@
 
 #include "framework/physics/solver_base/solver.h"
 #include "framework/math/petsc_utils/petsc_utils.h"
-#include "modules/common/diffusion_bndry.h"
+#include "modules/diffusion/boundary.h"
 #include "framework/utils/timer.h"
-#include "framework/math/unknown_manager/unknown_manager.h"
-#include "framework/mesh/mesh.h"
 #include <map>
+
+#include "framework/mesh/mesh.h"
 
 namespace opensn
 {
@@ -21,14 +21,15 @@ namespace diffusion
 {
 
 /**
- * DFEM diffusion solver
+ * CFEM diffusion solver
+ *
  */
-class DFEMSolver : public opensn::Solver
+class CFEMSolver : public opensn::Solver
 {
 public:
-  explicit DFEMSolver(const std::string& name);
-  explicit DFEMSolver(const InputParameters& params);
-  ~DFEMSolver() override;
+  explicit CFEMSolver(const std::string& name);
+  explicit CFEMSolver(const InputParameters& params);
+  ~CFEMSolver() override;
 
   void SetDCoefFunction(std::shared_ptr<ScalarSpatialMaterialFunction> function);
   void SetQExtFunction(std::shared_ptr<ScalarSpatialMaterialFunction> function);
@@ -38,6 +39,7 @@ public:
   void SetBoundaryOptions(const InputParameters& params);
 
   void Initialize() override;
+
   void Execute() override;
 
   /**
@@ -46,30 +48,6 @@ public:
   void UpdateFieldFunctions();
 
 private:
-  /**Still searching for a reference for this.
-   *
-   * For Polygons:
-   * Defined from paper  \n
-   * Turcksin B, Ragusa J, "Discontinuous diffusion synthetic acceleration
-   * for S_n transport on 2D arbitrary polygonal meshes", Journal of
-   * Computational Physics 274, pg 356-369, 2014.\n
-   * \n
-   * Nv = Number of vertices. If Nv <= 4 then the perimeter parameter
-   * should be replaced by edge length.*/
-  double HPerpendicular(const Cell& cell, unsigned int f);
-
-  /**
-   * Maps a face, in a discontinuous sense, using the spatial discretization.
-   */
-  int MapFaceNodeDisc(const Cell& cur_cell,
-                      const Cell& adj_cell,
-                      const std::vector<Vector3>& cc_node_locs,
-                      const std::vector<Vector3>& ac_node_locs,
-                      size_t ccf,
-                      size_t acf,
-                      size_t ccfi,
-                      double epsilon = 1.0e-12);
-
   typedef std::pair<opensn::diffusion::BoundaryType, std::vector<double>> BoundaryInfo;
   typedef std::map<std::string, BoundaryInfo> BoundaryPreferences;
 
@@ -79,8 +57,6 @@ private:
 
   size_t num_local_dofs_ = 0;
   size_t num_globl_dofs_ = 0;
-
-  std::vector<double> field_;
 
   /// approx solution
   Vec x_ = nullptr;
