+++ modules/diffusion/dfem_diffusion_solver.h
@@ -5,9 +5,10 @@
 
 #include "framework/physics/solver_base/solver.h"
 #include "framework/math/petsc_utils/petsc_utils.h"
+#include "modules/diffusion/boundary.h"
 #include "framework/utils/timer.h"
+#include "framework/math/unknown_manager/unknown_manager.h"
 #include "framework/mesh/mesh.h"
-#include "modules/common/diffusion_bndry.h"
 #include <map>
 
 namespace opensn
@@ -20,14 +21,14 @@ namespace diffusion
 {
 
 /**
- * FV diffusion solver
+ * DFEM diffusion solver
  */
-class FVSolver : public opensn::Solver
+class DFEMSolver : public opensn::Solver
 {
 public:
-  explicit FVSolver(const std::string& name);
-  explicit FVSolver(const InputParameters& params);
-  ~FVSolver() override;
+  explicit DFEMSolver(const std::string& name);
+  explicit DFEMSolver(const InputParameters& params);
+  ~DFEMSolver() override;
 
   void SetDCoefFunction(std::shared_ptr<ScalarSpatialMaterialFunction> function);
   void SetQExtFunction(std::shared_ptr<ScalarSpatialMaterialFunction> function);
@@ -36,30 +37,57 @@ public:
   void SetOptions(const InputParameters& params);
   void SetBoundaryOptions(const InputParameters& params);
 
-  // void Initialize() override;
   void Initialize() override;
   void Execute() override;
 
-  /**Updates the field functions with the latest data.*/
+  /**
+   * Updates the field functions with the latest data.
+   */
   void UpdateFieldFunctions();
 
 private:
+  /**Still searching for a reference for this.
+   *
+   * For Polygons:
+   * Defined from paper  \n
+   * Turcksin B, Ragusa J, "Discontinuous diffusion synthetic acceleration
+   * for S_n transport on 2D arbitrary polygonal meshes", Journal of
+   * Computational Physics 274, pg 356-369, 2014.\n
+   * \n
+   * Nv = Number of vertices. If Nv <= 4 then the perimeter parameter
+   * should be replaced by edge length.*/
+  double HPerpendicular(const Cell& cell, unsigned int f);
+
+  /**
+   * Maps a face, in a discontinuous sense, using the spatial discretization.
+   */
+  int MapFaceNodeDisc(const Cell& cur_cell,
+                      const Cell& adj_cell,
+                      const std::vector<Vector3>& cc_node_locs,
+                      const std::vector<Vector3>& ac_node_locs,
+                      size_t ccf,
+                      size_t acf,
+                      size_t ccfi,
+                      double epsilon = 1.0e-12);
+
   typedef std::pair<opensn::diffusion::BoundaryType, std::vector<double>> BoundaryInfo;
   typedef std::map<std::string, BoundaryInfo> BoundaryPreferences;
 
-  std::shared_ptr<MeshContinuum> grid_ptr_;
+  std::shared_ptr<MeshContinuum> grid_ptr_ = nullptr;
 
-  std::shared_ptr<SpatialDiscretization> sdm_ptr_;
+  std::shared_ptr<SpatialDiscretization> sdm_ptr_ = nullptr;
 
-  size_t num_local_dofs_;
-  size_t num_global_dofs_;
+  size_t num_local_dofs_ = 0;
+  size_t num_globl_dofs_ = 0;
+
+  std::vector<double> field_;
 
   /// approx solution
-  Vec x_;
+  Vec x_ = nullptr;
   /// RHS
-  Vec b_;
+  Vec b_ = nullptr;
   /// linear system matrix
-  Mat A_;
+  Mat A_ = nullptr;
 
   std::map<uint64_t, Boundary> boundaries_;
