+++ modules/diffusion/fv_diffusion_solver.cc
@@ -1,7 +1,7 @@
 // SPDX-FileCopyrightText: 2024 The OpenSn Authors <https://open-sn.github.io/opensn/>
 // SPDX-License-Identifier: MIT
 
-#include "modules/fv_diffusion/fv_diffusion_solver.h"
+#include "modules/diffusion/fv_diffusion_solver.h"
 #include "framework/runtime.h"
 #include "framework/object_factory.h"
 #include "framework/logging/log.h"
