+++ modules/diffusion/fv_diffusion_solver.h
@@ -5,11 +5,10 @@
 
 #include "framework/physics/solver_base/solver.h"
 #include "framework/math/petsc_utils/petsc_utils.h"
-#include "modules/common/diffusion_bndry.h"
 #include "framework/utils/timer.h"
-#include <map>
-
 #include "framework/mesh/mesh.h"
+#include "modules/diffusion/boundary.h"
+#include <map>
 
 namespace opensn
 {
@@ -21,15 +20,14 @@ namespace diffusion
 {
 
 /**
- * CFEM diffusion solver
- *
+ * FV diffusion solver
  */
-class CFEMSolver : public opensn::Solver
+class FVSolver : public opensn::Solver
 {
 public:
-  explicit CFEMSolver(const std::string& name);
-  explicit CFEMSolver(const InputParameters& params);
-  ~CFEMSolver() override;
+  explicit FVSolver(const std::string& name);
+  explicit FVSolver(const InputParameters& params);
+  ~FVSolver() override;
 
   void SetDCoefFunction(std::shared_ptr<ScalarSpatialMaterialFunction> function);
   void SetQExtFunction(std::shared_ptr<ScalarSpatialMaterialFunction> function);
@@ -38,32 +36,30 @@ public:
   void SetOptions(const InputParameters& params);
   void SetBoundaryOptions(const InputParameters& params);
 
+  // void Initialize() override;
   void Initialize() override;
-
   void Execute() override;
 
-  /**
-   * Updates the field functions with the latest data.
-   */
+  /**Updates the field functions with the latest data.*/
   void UpdateFieldFunctions();
 
 private:
   typedef std::pair<opensn::diffusion::BoundaryType, std::vector<double>> BoundaryInfo;
   typedef std::map<std::string, BoundaryInfo> BoundaryPreferences;
 
-  std::shared_ptr<MeshContinuum> grid_ptr_ = nullptr;
+  std::shared_ptr<MeshContinuum> grid_ptr_;
 
-  std::shared_ptr<SpatialDiscretization> sdm_ptr_ = nullptr;
+  std::shared_ptr<SpatialDiscretization> sdm_ptr_;
 
-  size_t num_local_dofs_ = 0;
-  size_t num_globl_dofs_ = 0;
+  size_t num_local_dofs_;
+  size_t num_global_dofs_;
 
   /// approx solution
-  Vec x_ = nullptr;
+  Vec x_;
   /// RHS
-  Vec b_ = nullptr;
+  Vec b_;
   /// linear system matrix
-  Mat A_ = nullptr;
+  Mat A_;
 
   std::map<uint64_t, Boundary> boundaries_;
